Refactor Inventory API and Consolidate Export Functionality
- Removed deprecated AssetsExportAPI and FindingsExportAPI, consolidating their functionalities into a new unified ExportAPI.
- Updated InventoryAPI to include the new export interface, enhancing the overall structure and usability.
- Deleted outdated documentation files for assets and findings export, while adding new documentation for the unified export API.
- Updated tests to reflect the changes in the export functionality and ensure comprehensive coverage.
