--- docs/api/tenableone/inventory/assets/schema.rst
@@ -1,1 +1,0 @@
-.. automodule:: tenable.tenableone.inventory.assets.schema
