--- docs/api/tenableone/inventory/assets_export/index.rst
@@ -1,1 +1,0 @@
-.. automodule:: tenable.tenableone.inventory.assets_export.api
