--- docs/api/tenableone/inventory/findings_export/index.rst
@@ -1,1 +1,0 @@
-.. automodule:: tenable.tenableone.inventory.findings_export.api
