--- docs/api/tenableone/inventory/findings_export/schema.rst
@@ -1,1 +1,0 @@
-.. automodule:: tenable.tenableone.inventory.export_schema
