--- tenable/tenableone/inventory/findings_export/__init__.py
@@ -1,1 +1,0 @@
-
