--- tests/tenableone/inventory/assets_export/__init__.py
@@ -1,1 +1,0 @@
-
